- optimize performance
- chatglm adapter TP support

before
```
concurrency: 256
elapsed_time: 306.702s

first token latency(s)(min, max, ave): 0.281, 10.159, 1.888
per-token latency(s) percentile(50, 75, 95, 99): [0.083, 0.088, 0.271, 0.444]

number of prompt tokens: 724392
number of completion tokens: 662960
token throughput (completion token): 2161.578 token/s
token throughput (prompt + completion token): 4523.455 token/s
RPS (request per second): 9.781 req/s
RPM (request per minute): 586.889 req/min
```

after
```
concurrency: 256
elapsed_time: 261.317s

first token latency(s)(min, max, ave): 0.203, 9.053, 1.658
per-token latency(s) percentile(50, 75, 95, 99): [0.069, 0.077, 0.242, 0.409]

number of prompt tokens: 724392
number of completion tokens: 662960
token throughput (completion token): 2536.999 token/s
token throughput (prompt + completion token): 5309.085 token/s
RPS (request per second): 11.480 req/s
RPM (request per minute): 688.820 req/min
```